Перенес коннект из withSignIn в сам компонент
