+++ src/components/sign-in/sign-in.jsx
@@ -1,5 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import {connect} from 'react-redux';
+import {Redirect} from 'react-router-dom';
+import {FieldsIds, AppRoutes} from '../../consts/consts.js';
+import {getAuthorizationStatusBoolean, getLoginStatusCode, getIsLoading} from '../../reducer/selectors.js';
+import {Operation as UserOperation, ActionCreator as UserActionCreator} from '../../reducer/user/user.js';
+import {Operation as MovieDataOperation, ActionCreator as MovieDataActionCreator} from '../../reducer/data/movies-data/movies-data.js';
+
 import Header from '../header/header.jsx';
 import Footer from '../footer/footer.jsx';
 import InputField from '../input-filed/input-field.jsx';
@@ -8,6 +15,24 @@ import withSignIn from '../../hocs/with-sign-in/with-sign-in.js';
 import UserProfile from '../user-profile/user-profile.jsx';
 
 const SignIn = (props) => {
+  const handleFormSubmit = (evt) => {
+    evt.preventDefault();
+    const {login, isFormValid} = props;
+    if (isFormValid) {
+      login({
+        email: props[FieldsIds.EMAIL_FIELD_ID].value,
+        password: props[FieldsIds.PASSWORD_FIELD_ID].value,
+      });
+    }
+  };
+
+  if (props.isAuthorized) {
+    props.getUserFavoriteList();
+    return (
+      <Redirect to={AppRoutes.MAIN_PAGE} />
+    );
+  }
+
   return (
     <div className="user-page">
 
@@ -22,7 +47,7 @@ const SignIn = (props) => {
           <p>We can’t recognize this email <br /> and password combination. Please try again.</p>
         </div>
         }
-        <form onSubmit={props.onFormSubmit} className="sign-in__form">
+        <form onSubmit={handleFormSubmit} className="sign-in__form">
           <div className="sign-in__fields">
             {
               props.signInFields.map((fld) => {
@@ -59,7 +84,6 @@ SignIn.propTypes = {
   loginStatusCode: PropTypes.number.isRequired,
   isAuthorized: PropTypes.bool.isRequired,
   onInputChange: PropTypes.func.isRequired,
-  onFormSubmit: PropTypes.func.isRequired,
   signInFields: PropTypes.arrayOf(
       PropTypes.exact({
         id: PropTypes.string.isRequired,
@@ -67,8 +91,28 @@ SignIn.propTypes = {
         type: PropTypes.string.isRequired,
         placeholder: PropTypes.string.isRequired,
       }).isRequired
-  )
+  ),
+  login: PropTypes.func.isRequired,
+  getUserFavoriteList: PropTypes.func.isRequired,
+  isFormValid: PropTypes.bool.isRequired,
 };
 
+const mapStateToProps = (state) => ({
+  isAuthorized: getAuthorizationStatusBoolean(state),
+  loginStatusCode: getLoginStatusCode(state),
+  isLoading: getIsLoading(state),
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  login(authData) {
+    dispatch(UserActionCreator.setAuthorizationProgress(true));
+    dispatch(UserOperation.login(authData));
+  },
+  getUserFavoriteList() {
+    dispatch(MovieDataActionCreator.fetchUserFavoriteList());
+    dispatch(MovieDataOperation.fetchUserFavoriteListData());
+  }
+});
+
 export {SignIn};
-export default withSignIn(SignIn);
+export default connect(mapStateToProps, mapDispatchToProps)(withSignIn(SignIn));
